Move util-related files into separated folder + refactor imports
