+++ handlers/admin/admin_coupons.py
@@ -6,7 +6,7 @@ from aiogram.fsm.state import State, StatesGroup
 from aiogram.types import InlineKeyboardButton
 from aiogram.utils.keyboard import InlineKeyboardBuilder
 
-from database import create_coupon, delete_coupon_from_db, get_all_coupons
+from utils.database import create_coupon, delete_coupon_from_db, get_all_coupons
 from filters.admin import IsAdminFilter
 from logger import logger
