+++ handlers/admin/admin_user_editor.py
@@ -9,7 +9,7 @@ from aiogram.types import CallbackQuery, InlineKeyboardButton
 from aiogram.utils.keyboard import InlineKeyboardBuilder
 
 from config import TOTAL_GB
-from database import get_client_id_by_email, get_servers_from_db, restore_trial, update_key_expiry
+from utils.database import get_client_id_by_email, get_servers_from_db, restore_trial, update_key_expiry
 from filters.admin import IsAdminFilter
 from handlers.keys.key_utils import delete_key_from_cluster, delete_key_from_db, renew_key_in_cluster
 from handlers.utils import sanitize_key_name
