+++ handlers/keys/key_utils.py
@@ -3,9 +3,9 @@ import asyncio
 import asyncpg
 from py3xui import AsyncApi
 
-from client import add_client, delete_client, extend_client_key
+from utils.client import add_client, delete_client, extend_client_key
 from config import ADMIN_PASSWORD, ADMIN_USERNAME, DATABASE_URL, TOTAL_GB
-from database import get_servers_from_db
+from utils.database import get_servers_from_db
 from logger import logger
