+++ handlers/keys/keys.py
@@ -9,7 +9,7 @@ from aiogram.types import BufferedInputFile, InlineKeyboardButton
 from aiogram.utils.keyboard import InlineKeyboardBuilder
 
 from config import CONNECT_ANDROID, CONNECT_IOS, DOWNLOAD_ANDROID, DOWNLOAD_IOS, PUBLIC_LINK, RENEWAL_PLANS, TOTAL_GB
-from database import delete_key, get_balance, get_servers_from_db, store_key, update_balance, update_key_expiry
+from utils.database import delete_key, get_balance, get_servers_from_db, store_key, update_balance, update_key_expiry
 from handlers.keys.key_utils import (
     delete_key_from_cluster,
     delete_key_from_db,
