+++ handlers/keys/trial_key.py
@@ -5,9 +5,9 @@ import uuid
 
 from py3xui import AsyncApi
 
-from client import add_client
+from utils.client import add_client
 from config import ADMIN_PASSWORD, ADMIN_USERNAME, PUBLIC_LINK, TOTAL_GB, TRIAL_TIME
-from database import get_servers_from_db, store_key, use_trial
+from utils.database import get_servers_from_db, store_key, use_trial
 from handlers.texts import INSTRUCTIONS
 from handlers.utils import generate_random_email, get_least_loaded_cluster
