+++ handlers/payments/freekassa_pay.py
@@ -13,7 +13,7 @@ from aiohttp import web
 import requests
 
 from config import FREEKASSA_API_KEY, FREEKASSA_SHOP_ID
-from database import add_payment, update_balance
+from utils.database import add_payment, update_balance
 from handlers.payments.utils import send_payment_success_notification
 from handlers.texts import PAYMENT_OPTIONS
