+++ handlers/payments/robokassa_pay.py
@@ -10,7 +10,7 @@ from aiohttp import web
 from robokassa import HashAlgorithm, Robokassa
 
 from config import ROBOKASSA_ENABLE, ROBOKASSA_LOGIN, ROBOKASSA_PASSWORD1, ROBOKASSA_PASSWORD2, ROBOKASSA_TEST_MODE
-from database import add_connection, add_payment, check_connection_exists, get_key_count, update_balance
+from utils.database import add_connection, add_payment, check_connection_exists, get_key_count, update_balance
 from handlers.payments.utils import send_payment_success_notification
 from handlers.texts import PAYMENT_OPTIONS
 from logger import logger
