+++ main.py
@@ -4,7 +4,7 @@ import signal
 from aiogram.webhook.aiohttp_server import SimpleRequestHandler, setup_application
 from aiohttp import web
 
-from backup import backup_database
+from utils.backup import backup_database
 from bot import bot, dp, router
 from config import (
     BACKUP_TIME,
@@ -20,7 +20,7 @@ from config import (
     WEBHOOK_URL,
     YOOKASSA_ENABLE,
 )
-from database import init_db
+from utils.database import init_db
 from handlers.keys.subscriptions import handle_new_subscription, handle_old_subscription
 from handlers.notifications import notify_expiring_keys
 from handlers.payments.cryprobot_pay import cryptobot_webhook
@@ -28,7 +28,7 @@ from handlers.payments.freekassa_pay import freekassa_webhook
 from handlers.payments.robokassa_pay import robokassa_webhook
 from handlers.payments.yookassa_pay import yookassa_webhook
 from logger import logger
-from servers import check_servers, sync_servers_with_db
+from utils.servers import check_servers, sync_servers_with_db
 
 
 async def periodic_notifications():
